demo-app/requirements.txt
Flask==3.0.3
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-flask==0.46b0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-instrumentation-requests==0.46b0
requests==2.32.3
pyroscope-io==0.8.11
